Split cli.py into multiple files, because `dxr deploy` is going to be quite long.
